Move some schema extraction login to cli-build
