+++ packages/@sanity/cli-build/package.json
@@ -53,8 +53,10 @@
   },
   "dependencies": {
     "@sanity/cli-core": "workspace:^",
+    "@sanity/schema": "catalog:",
     "read-package-up": "catalog:",
-    "semver": "^7.7.4"
+    "semver": "^7.7.4",
+    "zod": "catalog:"
   },
   "devDependencies": {
     "@eslint/compat": "catalog:",
@@ -72,6 +74,7 @@
     "@vitest/coverage-istanbul": "catalog:",
     "eslint": "catalog:",
     "publint": "catalog:",
+    "sanity": "catalog:",
     "typescript": "catalog:",
     "vitest": "catalog:"
   },
