+++ packages/@sanity/cli-build/src/_exports/_internal.ts
@@ -3,5 +3,8 @@ export {checkStudioDependencyVersions} from '../actions/build/checkStudioDepende
 export {generateWebManifest} from '../actions/build/generateWebManifest.js'
 export {getDefaultFaviconsPath, writeFavicons} from '../actions/build/writeFavicons.js'
 export {type ExtractOptions, getExtractOptions} from '../actions/schema/getExtractOptions.js'
+export {runSchemaExtraction} from '../actions/schema/runSchemaExtraction.js'
+export {type ExtractSchemaWorkerError} from '../actions/schema/types.js'
+export {extractValidationFromSchemaError} from '../actions/schema/utils/extractValidationFromSchemaError.js'
 export {SchemaExtractionError} from '../actions/schema/utils/SchemaExtractionError.js'
 export {copyDir} from '../util/copyDir.js'
