+++ packages/@sanity/cli-build/src/actions/schema/utils/extractValidationFromSchemaError.ts
