+++ packages/@sanity/cli-core/src/_exports/index.ts
@@ -40,6 +40,7 @@ export * from '../util/environment/mockBrowserEnvironment.js'
 export * from '../util/getLocalPackageVersion.js'
 export * from '../util/getSanityEnvVar.js'
 export * from '../util/getSanityUrl.js'
+export * from '../util/getWorkspace.js'
 export * from '../util/importModule.js'
 export * from '../util/isCi.js'
 export * from '../util/isInteractive.js'
