+++ packages/@sanity/cli/src/actions/manifest/extractManifest.ts
@@ -1,8 +1,7 @@
-import {SchemaExtractionError} from '@sanity/cli-build/_internal'
+import {type ExtractSchemaWorkerError, SchemaExtractionError} from '@sanity/cli-build/_internal'
 import {findProjectRoot, getTimer, studioWorkerTask} from '@sanity/cli-core'
 import {spinner} from '@sanity/cli-core/ux'
 
-import {type ExtractSchemaWorkerError} from '../schema/types.js'
 import {manifestDebug} from './debug.js'
 import {type CreateWorkspaceManifest, type ExtractManifestWorkerData} from './types'
 import {writeManifestFile} from './writeManifestFile.js'
