+++ packages/@sanity/cli/src/actions/manifest/extractManifest.worker.ts
@@ -1,8 +1,8 @@
 import {isMainThread, parentPort, workerData} from 'node:worker_threads'
 
+import {extractValidationFromSchemaError} from '@sanity/cli-build/_internal'
 import {getStudioWorkspaces, subdebug} from '@sanity/cli-core'
 
-import {extractValidationFromSchemaError} from '../schema/utils/extractValidationFromSchemaError.js'
 import {extractWorkspaceManifest} from './extractWorkspaceManifest.js'
 import {extractManifestWorkerData} from './types.js'
