+++ packages/@sanity/cli/src/actions/schema/extractSchema.ts
@@ -1,11 +1,14 @@
 import {exit} from '@oclif/core/errors'
-import {type ExtractOptions, SchemaExtractionError} from '@sanity/cli-build/_internal'
+import {
+  type ExtractOptions,
+  runSchemaExtraction,
+  SchemaExtractionError,
+} from '@sanity/cli-build/_internal'
 import {getCliTelemetry, type Output} from '@sanity/cli-core'
 import {spinner} from '@sanity/cli-core/ux'
 
 import {SchemaExtractedTrace} from '../../telemetry/extractSchema.telemetry.js'
 import {formatSchemaValidation} from './formatSchemaValidation.js'
-import {runSchemaExtraction} from './runSchemaExtraction.js'
 import {schemasExtractDebug} from './utils/debug.js'
 
 interface ExtractSchemaActionOptions {
