+++ packages/@sanity/cli/src/actions/schema/vite/plugin-schema-extraction.ts
@@ -1,6 +1,6 @@
 import path, {isAbsolute} from 'node:path'
 
-import {SchemaExtractionError} from '@sanity/cli-build/_internal'
+import {runSchemaExtraction, SchemaExtractionError} from '@sanity/cli-build/_internal'
 import {type CLITelemetryStore} from '@sanity/cli-core'
 import {logSymbols} from '@sanity/cli-core/ux'
 import debounce from 'lodash-es/debounce.js'
@@ -14,7 +14,6 @@ import {
 } from '../../../telemetry/extractSchema.telemetry.js'
 import {formatSchemaValidation} from '../formatSchemaValidation.js'
 import {createSchemaPatternMatcher} from '../matchSchemaPattern.js'
-import {runSchemaExtraction} from '../runSchemaExtraction.js'
 
 /**
  * Default glob patterns to watch for schema changes.
